windows

+++ src/raw_parse.rs
@@ -2179,7 +2179,12 @@ unsafe fn convert_close_portal_stmt(cps: &bindings_raw::ClosePortalStmt) -> prot
 }
 
 unsafe fn convert_fetch_stmt(fs: &bindings_raw::FetchStmt) -> protobuf::FetchStmt {
-    protobuf::FetchStmt { direction: fs.direction as i32 + 1, how_many: fs.howMany, portalname: convert_c_string(fs.portalname), ismove: fs.ismove }
+    protobuf::FetchStmt {
+        direction: fs.direction as i32 + 1,
+        how_many: fs.howMany as i64,
+        portalname: convert_c_string(fs.portalname),
+        ismove: fs.ismove,
+    }
 }
 
 unsafe fn convert_declare_cursor_stmt(dcs: &bindings_raw::DeclareCursorStmt) -> protobuf::DeclareCursorStmt {
